Support use on 32-bit architectures
The previous version of the code would normally fail when used on a
32-bit architecture with a message "panic: unaligned 64-bit atomic
operation". It turns out that `atomic` requires loads and stores on
64-bit numbers that are 64-bit aligned, which doesn't happen by
default on 32-bit architectures.

Fortunately, the `atomic.Uint64` type is guaranteed to always be
64-bit aligned. (See https://pkg.go.dev/sync/atomic#pkg-note-BUG). So
we can instead just use these instead of plain uint64 values. This
commit changes all the unit64 values that we atomically load or store
into `atomic.Uint64`.

This code has been tested on a 32-bit ARMv6 processor (and now works!)